+++ packages/backend/src/services/HealthService/HealthService.ts
@@ -98,6 +98,7 @@ export class HealthService {
             hasDbtSemanticLayer:
                 !!process.env.DBT_CLOUD_ENVIRONMENT_ID &&
                 !!process.env.DBT_CLOUD_BEARER_TOKEN,
+            hasGroups: this.lightdashConfig.groups.enabled,
         };
     }