+++ packages/common/src/types/projects.ts
@@ -44,6 +44,7 @@ export type CreateBigqueryCredentials = {
     timeoutSeconds: number | undefined;
     priority: 'interactive' | 'batch' | undefined;
     keyfileContents: Record<string, string>;
+    requireUserCredentials?: boolean;
     retries: number | undefined;
     location: string | undefined;
     maximumBytesBilled: number | undefined;
@@ -72,6 +73,7 @@ export type CreateDatabricksCredentials = {
     serverHostName: string;
     httpPath: string;
     personalAccessToken: string;
+    requireUserCredentials?: boolean;
     startOfWeek?: WeekDay | null;
 };
 export type DatabricksCredentials = Omit<
@@ -83,6 +85,7 @@ export type CreatePostgresCredentials = SshTunnelConfiguration & {
     host: string;
     user: string;
     password: string;
+    requireUserCredentials?: boolean;
     port: number;
     dbname: string;
     schema: string;
@@ -102,6 +105,7 @@ export type CreateTrinoCredentials = {
     host: string;
     user: string;
     password: string;
+    requireUserCredentials?: boolean;
     port: number;
     dbname: string;
     schema: string;
@@ -117,6 +121,7 @@ export type CreateRedshiftCredentials = SshTunnelConfiguration & {
     host: string;
     user: string;
     password: string;
+    requireUserCredentials?: boolean;
     port: number;
     dbname: string;
     schema: string;
@@ -135,6 +140,7 @@ export type CreateSnowflakeCredentials = {
     account: string;
     user: string;
     password?: string;
+    requireUserCredentials?: boolean;
     privateKey?: string;
     privateKeyPass?: string;
     role?: string;